Fix the broken log_destination = syslog functionality.
Fixing the logging to the syslog destination, which got broken by the
PGPOOL SET/SHOW command commit, and also enhancing the log_destination
configuration parameter to be assigned with the comma separated list of multiple
destinations for the Pgpool-II log. Now, after this commit log_destination can
be set to any combination of 'syslog' and 'stderr' log destinations.