Check if "milestone close" maybe received or not before checking pending data.
pool_push_pending_data() checks fd of backend if buffers are empty
with select(2) timers is 0.  This may loose pending data which is not
processed in backend yet. To minimize the chance loosing pending data,
check the previously received data is "milestone close" (close
complete) or not. If it was not the miles-stone close, set the timer
to "wait forever" because we can be sure that the miles-stone close
will come later. Per pgpool-hackers: 2103.
http://www.pgpool.net/pipermail/pgpool-hackers/2017-March/002103.html