Fix issue with bug271.
When an INSERT is executed in a transaction, a "writing_transaction"
is set. The flag is checked by a succeeding SELECT to decide if
parse_before_bind (which resend the parse message to primary node if
it is executed on the standby) needs to be executed.  The flag is
reset when a command complete message for BEGIN is arrived.  Problem
is, it is possible that the command complete message could arrive
before the bind for the SELECT is executed, and at this point the flag
is already cleared.

Solution is, do not reset the flag when the command complete message
arrives if in streaming replication mode.