Fix Execute() not to set writing transaction flag in an inappropriate case.
Execute() mistakenly set the writing transaction flag even if a write
query is issued (in particular case, "BEGIN"). The reason why it did
not found before was, the code path only works when the transaction
state is 'T' (in transaction). Before it was only set at
CommandComplete but previous commit set the state at Execute().
This leads to a SELECT not being load balanced.

Also add some useful debugging statements.

See [pgpool-hackers: 2155] for more details.