Fix failing in re-syncing primary and standby when "ready for query" received.
After sending a sync message to all backend nodes, eventually a ready
for query message comes from all nodes. So read_data_from_backend()
syncs all nodes by reading and discarding messages until "ready for
query" is coming. Before it checks whether data is ready or not before
reading from socket with timeout to 0 when data input buffer is
empty. However this is not reliable since the data may not be ready at
the timing. Moreover this could lead to a kind mismatch error since
the re-syncing process is aborted. Now it is safe to assume that data
must be eventually ready, so the timeout is not necessary anymore.

[See pgpool-hacker: 2167] for more details.