Fix the case when duplicate statement remains.
By using parse_before_bind, it is possible a parsed statement remains
on the load balance node even if explicit close is issued because the
close is redirected to the primary node. Fix is, issue close to both
primary and load balance node anytime. This sounds is a little bit
over killing but I think there's no way to remember the statement
remains on a load balance node.

Also clean up some ifdef NOT_USED garbage.