Fix for 0000299: Errors on the reloading of configuration
Pgpool-II should load the hba_file after evaluating all the command line
options at startup.
It is wrong to load hba file before evaluating all the command line arguments,
because when the Pgpool-II is invoked with reload or stop options, all it
needs to do is to signal the running Pgpool-II and exit, So loading hba file
for pgpool reload/stop is not only unnecessary but It also emits the
misleading WARNING message of "failed while loading hba configuration"
when alternative path for hba file was used.