Fix for [pgpool-general: 5621] pgpool 3.6.4 failover
Failover() function should be executed with health check alarm disabled

The reason is the find primary node and health check uses the same function to
make a connection to the backends, and the backend connect function has a built
in logic to bail out with the timeout error when the health check alarm occurs.
Now the problem is if the failover() is executed while the health check alarm is
still running, Their is a strong possibility that the find primary node function
may find the health check timer already expired and exit without locating the
new primary node.

The commit also fixes a spelling mistake in the function name and
enum definition. The typo was identified by Ishii-San