Fix for bug in watchdog where sometime failover is not reliably performed
Currently watchdog process only considers the node's watchdog state before
deciding if it can handle the failover and failover-locking requests. But the
problem with this technique is that, for the instance when the node has
announced itself as a master/coordinator of the cluster and is waiting for
the standby nodes to recognise it as a Master node. For that period of time
the watchdog state of the node is Master/coordinator, but it is yet not fully
capable of handling the failover and failover-locking requests.
So sometimes this leads to a situation where a failover is not reliably
performed in case when the failover request arrives while watchdog cluster
is in the process of electing a leader.

The fix for the above situation is to make sure that the node has fully acquired
the Master status before accepting the failover and failover-locking requests.